backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
gitpython==3.1.40
bandit==1.7.5
semgrep==1.34.0
google-generativeai==0.3.1
python-jose==3.3.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
pytest==7.4.3
httpx==0.25.1 